pyproject.toml
[project]
name = "ai-102"
version = "0.1.0"
description = "ai-102 playground"
authors = [
    {name = "dhruv9586",email = "dhruv9586@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openai (>=1.107.2,<2.0.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "requests (>=2.32.5,<3.0.0)",
    "azure-search-documents (>=11.5.3,<12.0.0)",
    "azure-ai-contentsafety (>=1.0.0,<2.0.0)",
    "azure-ai-vision-imageanalysis (>=1.0.0,<2.0.0)",
    "azure-ai-vision-face (>=1.0.0b2,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

package-mode = false